requirements.txt
discord.py==1.5.1
gTTS==2.1.1
PyAudio==0.2.11
